update crud ok

+++ routes/categorias.js
@@ -29,14 +29,24 @@ router.put('/:putIdCategoria', async (req, res) => {
     });
 })
 
-router.delete('/', async (req, res) => {
+router.delete('/:deleteIdCategoria', async (req, res) => {
     const {deleteIdCategoria} = req.params;
 
-    const deleteCategoria = await prisma.categoria.delete({
+    const consultaLivros = await prisma.livros.findMany({
         where: {
             id_categoria: deleteIdCategoria,
         }
     })
+
+    if (consultaLivros == '') {
+        const deleteCategoria = await prisma.categoria.delete({
+            where: {
+                id_categoria: Number(deleteIdCategoria),
+            }
+        })
+    } else {
+        
+    }
 })
 
 module.exports = router;
